+++ Components/Optimizers/PreconditionedStochasticGradientDescent/elxPreconditionedStochasticGradientDescent.hxx
@@ -164,7 +164,7 @@ PreconditionedStochasticGradientDescent<TElastix>::BeforeEachResolution()
      * This is a rather crude rule of thumb, which seems to work in practice.
      */
     this->m_NumberOfJacobianMeasurements =
-      std::max(static_cast<unsigned int>(5000), static_cast<unsigned int>(2 * numberOfParameters));
+      std::max(5000U, static_cast<unsigned int>(2 * numberOfParameters));
     configuration.ReadParameter(
       this->m_NumberOfJacobianMeasurements, "NumberOfJacobianMeasurements", this->GetComponentLabel(), level, 0);
 
@@ -179,7 +179,7 @@ PreconditionedStochasticGradientDescent<TElastix>::BeforeEachResolution()
      * This is a rather crude rule of thumb, which seems to work in practice.
      */
     this->m_NumberOfSamplesForPrecondition =
-      std::max(static_cast<unsigned int>(1000), static_cast<unsigned int>(numberOfParameters));
+      std::max(1000U, static_cast<unsigned int>(numberOfParameters));
     configuration.ReadParameter(
       this->m_NumberOfSamplesForPrecondition, "NumberOfSamplesForPrecondition", this->GetComponentLabel(), level, 0);
